# openapi/media-svc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/share-items/user/{userSub}/playlist/{playlistId}": {
      "post": {
        "operationId": "ShareItemController_sharePlaylist",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userSub",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistShareItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/user/{userSub}/media-item/{mediaId}": {
      "post": {
        "operationId": "ShareItemController_shareMediaItem",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userSub",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaShareItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-by-user": {
      "get": {
        "operationId": "ShareItemController_findItemsSharedByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareItemsDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-by-user/media-items": {
      "get": {
        "operationId": "ShareItemController_findMediaItemsSharedByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-by-user/playlists": {
      "get": {
        "operationId": "ShareItemController_findPlaylistsSharedByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-with-user": {
      "get": {
        "operationId": "ShareItemController_findItemsSharedWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItemsDto"
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-with-user/media-items": {
      "get": {
        "operationId": "ShareItemController_findMediaItemsSharedWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/shared-with-user/playlists": {
      "get": {
        "operationId": "ShareItemController_findPlaylistsSharedWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/{shareId}": {
      "get": {
        "operationId": "ShareItemController_findShareItem",
        "parameters": [
          {
            "name": "shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ShareItemController_removeShareItem",
        "parameters": [
          {
            "name": "shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/read/{shareId}": {
      "post": {
        "operationId": "ShareItemController_readShareItem",
        "parameters": [
          {
            "name": "shareId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/unshare-all-items": {
      "post": {
        "operationId": "ShareItemController_removeAllShareItems",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareItemsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/share-items/unshare-all-by-user-id": {
      "post": {
        "operationId": "ShareItemController_removeShareItemAllByUserId",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareItemsByUserIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "share-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Media Service",
    "description": "Media Service",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local"
    },
    {
      "url": "https://media-api.dev.afehrpt.com",
      "description": "staging"
    },
    {
      "url": "https://media-api.afehrpt.com",
      "description": "production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TagKeyValue": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "value"
        ]
      },
...
      "MediaVisibilityType": {
        "type": "string",
        "enum": [
          "private",
          "shared",
          "subscription",
          "public"
        ]
      },
...
      "MediaItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "key": {
            "type": "string",
            "example": "hk5NOHrpKvyqlnACOMld8DQb3y2Ta4Zu8fS3IzNiGVPeuhqEv1NwD8dTLQXkUrbD6nVHJLew5Y0u3Q5LFWdeEuTOQUhHP06jivSD09iQaQi8VCJQVIpbxApoDQnyA",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "7mvTNZX6wYVJuQUrunHlpoUjqrUXziIPq3n5xgyCFdm53NJO2DluqgEILWyzwGuY91U4EVTC5qzq9mFjhhaLuOwc6k9KXwBR1gfuvS1du6H3vB8Cn8ROFhPHni1YS",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "hBKzoS8DyCHy6ybVnUOihbCGNArHs6utlRkKIwnujs05LDlG75cMjn8IGpIMFDtdjIBnF7SVSJXyKUFJoXwwJp1U5lIE8g2aCpKKew5bN2IHxSCLRuOHlzY8xwS45",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "KKaZ0UaSY3K8YhuzwcahA70SX9TUchbiy38JZPQWFmZf3azg23c7A6JY6qIMqhpKxr5PGyAuP1kfgPifYrk7vBccyAEQexXXDN0KPW54EN2rAwtTuExfF7yD9Ubh1JiSQmJksYpbYdbTJDZJJhowLq8UnJfpEa7o9qWZlV7avZlcjWc7x62UIbGHx2bVTe3NhSLJExdc1Eik5BLaKwTFy7WFbSy8NcGiZA3I0hUflhoggj8FUCJ59XWW51fRiB3TrDrGdPhrY3BF8kWlaS1Ow3pOgffXdp0FVtRLpX1khkxx"
          },
          "description": {
            "type": "string",
            "example": "1AMrpE8mixH5tnZ7wLRAULuFzeUqNu4EeZohF1v1a7Z8nsxuADX0ilcfXV3dGjDcJXrS2ZavUKliDte9tZHF8b1tT0EdjNARdJxIAT5Z1slPVI0uexAwY1jOvXB0RT0DuULF4Bquma5TH9FAROI0Afi1iN5J4CuknXaMpU2kmLMHr1kDsmlPflOW4DjPbjRBL2ecG4DtDiArZ3zPaE16R1vjCkDaEg1WZ1hQufTX1xjSKRolDRK0GoxqrckyOJ79nCflkiDyzp8aFqZdTmjK3lZBfjIC19IkfhV2gN9RDHeV"
          },
          "uri": {
            "type": "string",
            "example": "jIQSOVTSCbQGZEIVkCWv5W5zhxsfzY8jBj98C3zb2GsAboCJ3MdmMh0P6A9L2zUyDqsmsozU9A84s3Sm61SGBCpHgOA78IiLU3HVmQ4TeIcEG0gGGtrkrW3wx7rwZuZclVW50QmVRUr9hmTmNiqTifJmsf5dDlRDIVgxxzwF4VaG53COK4zItVbvdzljUWoPfIaBVxRpVTdTkMzw2B3GjHupcUkdbQMqraKKYhjaUXkMSIdWeY4lzh131sZquVg3uSoE2qteYOIbP49oTuESp3ZoZpNwYzsZMjL178UbygdP"
          },
          "imageSrc": {
            "type": "string",
            "example": "7TJuQYS4yiKTFVNBygRTr72pCeIkrFFtEFSf8aBYE6ymk4ivX6yMescCepSgKJmifiZEBZiwMgsVRFfZ5Anm3EkTly2TpCU65FeFJNlRNkufi4lRPbAHzqxWAfm0f6nlFdH9ha1i8n9s0S54q6P0ZIoilFn4DjoEfqm6ODwPuuV77zqQkE648v4mjsq1poPWx7V3bbAo2pl6LWZICEP1nc9q5zBK440vV3liahgrR26lbw4QXAosizs4jUZechUo7hjju4zy5SLfJc28jzUFWl0zpwEnxC5v64vzIjMUGnRC"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared",
              "subscription",
              "public"
            ]
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "Zu2uSsVttocQw7iuPU418qu1Y5P4rI8Ch9QIyYtb0OO4qAefJgtKQowYLNpTpktezO4pkdkDcjPB6EfeLHhs4dbTZaRhpfJDzWOBEiVHt8gyOHwD2qDSBlvA29egG",
            "maxLength": 255,
            "minLength": 0
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "key",
          "userId",
          "title",
          "description",
          "uri",
          "visibility",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
...
      "PlaylistItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "playlistId": {
            "type": "string",
            "example": "NnkcrCJVUyKmQwL7jXEOoZUFWhxnrl3SxgIAZRYKvn09iD0mbNZ1HoezLaCwp4y5xmT6D5zEWUNBwHFpbR19YcWRwx0ScEXWBwiWsANkuQ22giTz4cF58AoO5frLf",
            "maxLength": 255,
            "minLength": 0
          },
          "mediaId": {
            "type": "string",
            "example": "MGOCQGdFerefknDcQ8yRozDq3sGooHecTfqvFnajScSHuYDGamNkh75nyQxpv0XhIU0hKavMo1p5Pq1jz7xwep74ujrcExYmVtC5EGOTSNobA233SaIhTHwuos3cL",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "7XhGLLQu0WVGeUbpNEsixGKOBtAGlzpK8CYyrqImaJKTzQ9dENqZkzCACklvIRWIm5S2egxlShAp8FJQrg0zmZzeWsaCQSKXLM7af76vSUD94WBvQrNxqLfzeheWG",
            "maxLength": 255,
            "minLength": 0
          },
          "createdBy": {
            "type": "string",
            "example": "4mJR8CdUC1LSdw7Cb3S8G2egMk7cwiAAWTfcMOo4oBWwSM9lZqva5iAuOPZDHHipBtVBfeCzY2xb3H47b695PCOFNCInO75fs2v7AHfWLL4F3JRG0BdiK6wAYaYQJ",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "rqODXvQmWUlupq4mLDzH0rCBPIJbMTagfj1lZ576zLISWDB4P11FIwJ8wzcmOv1LS9Ca8ZK8Cw5KsEIYolapMxwSk4LVtVcyACQa30ox7QW0TpUoyZ7BjjZA18NHP",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "Bd6aprPPaKMfZYIVLFOXsh7ruDifv3beRmzzNq0ARFqzYIj7Dp0YCgqYjSmIFEMnt7z7vk4JRH6UAg6om6Mvv7kPOmI7cGSYOMWE9nvLl2pXifV9Xb14frbrslekfeh9Caw5SwHOdFXpjjfjKgnoD45s531s5bTyLdCNKfiFmShCPUuJX4O6XbWTJwuQxH4sM3ipC1ZH8uXSzEv9wJdjdYvcGaJst58WmYxqx5DKlYUCM0JFsspAcOMkf4tLLGeRQEXaLpVjlBGX3fpxJH6S83twdYxv0zu5JvJmDifwFQjB"
          },
          "description": {
            "type": "string",
            "example": "7EaR2K3jjhuxmftuIYGyAEBzQj4oZ5HVxcwsCPPYT3Kd0ElvR93dU0u3dcqPqd8hGsV6Vvp998w3qDFn2UtTX73IXv1rn4CqrypsE1iu8MrWiNm2I2tlmodYRiDPKKfhonGyH7GY4ybQeDI1Uyxl97zmGd4BhZG3bLocglJUsxkmm1RQzGhaw6tJ7eY94SWGfZXiIM8TwOwIJThf7aCxyZ9ZXcowkr0S8DYns4ksOhiNiGq8WX1w6oZ7BSqDR92IzVfUlhgWDVDg0holBuqKb27mScDZjcayyyxHFGGy43eG"
          },
          "uri": {
            "type": "string",
            "example": "QPF8U2gIxmKnTNkRNhEbRfLEg4eKdq0T3so7qOTlbTAfrXU0r5NSO1fwkNU4vwh0Xkv7K4Qfm7aCFA86PH1u9xclkoO7G68FKvtu1l9EvXPMFCbMFCibo0SjlDyVp6iZth02cjNW8wCEeCjFEf95BcaOqNEcy9aDxeXQ4sxvBViqyzbYBYTaxBe9jwcLZxLRQIAEAenq27zIQ5DHTqZetXyX9v0jRu7Mijd8KF3IwJpMvafGBNoyMDpBAgkGlbBKB4UdGFfzo4KVgGDcSuTu5wPnuLNKzQqtVm1Ji1VaO6m9"
          },
          "imageSrc": {
            "type": "string",
            "example": "6MQdF7GEmKJnmK4fDNcJr7tfNdyij4VVXM3ePXRLD6T020MIeryeXAeTBsAQC6FJHL6EdWToJZYLQq6LaZx1seX9AEd3izEFgYwctZ5ou8f0lJ6H1oywMsmYJxmPIV1n0Nz2Ok7syr8BByZHW1pl3fcW0r242cg1SrwNj5kU2SJKH9SNf5XMpszVViYB2yQBp2MBXN11AGauJk8f9ycjasXfmNmaWbBQcdHwIUIXvAL43TSeo6WKt6e4busA0OPTdaMsf5nIHlCSmZjsS7m2suMrC5CmIhZG0c8YNzdZovZX"
          },
          "sortIndex": {
            "type": "number"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "$ref": "#/components/schemas/MediaVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "playlistId",
          "mediaId",
          "userId",
          "createdBy",
          "title",
          "description",
          "uri",
          "visibility",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
      "PlaylistVisibilityType": {
        "type": "string",
        "enum": [
          "private",
          "shared",
          "subscription",
          "public"
        ]
      },
...
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "cloneOf": {
            "type": "string",
            "example": "6881ed3b46854b3ebbade53b",
            "readOnly": false
          },
          "title": {
            "type": "string",
            "example": "S9keYXhfXP3fAT9H97JLufj2OF0M04CXnDtFWdTbiqZyN0dDN8n4HhgsoKqR8fDpB3uTgJ20nzQ2CA0A8rgMgUPHpzbhs9qM9Iwz204gKQbyvK2mNmnnKtuIEeE5e",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "example": "xa2hOk4q1oxSJPCbFfc2fhkvviS3qVN5YPceFvYcGBikCjmbKAfA06gdenSINzLbFCAtMudPwSrUlEBY8oRs5SMGEQQQ0AM0xxNWc1GhjDhXt7p56Kwtc6jnJTUdI8xmnvBlAo9x70UR70aNN7midrvStL4HLxt8uFpsRk2jx7VbnSJxyL2ihz04jBTv3F3pacrVlG5PPfAnCA6I9GIhXufQMRgh7WlGTvbDsep7HX0YJtg4kr4MPOegWF3IJG0bj1MNW1WaLQUnDak5czjFiWumA2eBsHJTNZbHPQYZAcaG"
          },
          "imageSrc": {
            "type": "string",
            "example": "Z5PzctoB54ep1AjFDPxTG8AMUfbuCm8fvwarnMBJz1nL2uHHmq9fJAPSYpmHcqKxOOwYWI0cETjplj0Z0Jp0Og9TkQfP8myV2fqUfrf7vqmDaPZMPBidHWa0yoqZe6dO6zKNLXTpDrcMMwRp2moikPV2gp8KxLPpu2thRpSHRC0LzwoXyDcPFJFdr1qaKxT2p6e1aQfBfKGNh8kBZh23TlNdE6WS08hoKY2UwsjFEpIWBu13U1dL5v5eRhLAPbKLU6CsoTav1Aua9RJgQl4pBdxYUHyrLftzm1F58Il82N3T"
          },
          "visibility": {
            "$ref": "#/components/schemas/PlaylistVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "mediaItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemDto"
            }
          },
          "playlistItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistItemDto"
            }
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "title",
          "description",
          "imageSrc",
          "visibility",
          "mediaItems",
          "playlistItems",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
...
      "PlaylistShareItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "shareId": {
            "type": "string"
          },
          "mediaId": {
            "type": "string"
          },
          "playlistId": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "shareId",
          "mediaId",
          "playlistId"
        ]
      },
      "MediaShareItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "shareId": {
            "type": "string"
          },
          "mediaId": {
            "type": "string"
          },
          "playlistId": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "shareId",
          "mediaId",
          "playlistId"
        ]
      },
      "ShareItemsDto": {
        "type": "object",
        "properties": {
          "mediaItems": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaShareItemDto"
              }
            ]
          },
          "playlists": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PlaylistShareItemDto"
              }
            ]
          }
        },
        "required": [
          "mediaItems",
          "playlists"
        ]
      },
      "ShareItem": {
        "type": "object",
        "properties": {}
      },
      "ShareItemsByUserIdDto": {
        "type": "object",
        "properties": {
          "shareItemByUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "shareItemByUserIds"
        ]
      }
    }
  }
}